Finished button class
It's really good. Don't mind if I pat myself on the back for this one. Really proud of the end result.
In all seriousness, I finished the button class, and left it in menutest.py. In future I might want other UI elements that I will just stick in the same file. Now that I know how to utalize inputManager better, It should be easy to add stuff.
Also deleted buttontest.py files